Modified applications.hh (converted weights from int to double) and minors appearence changes

+++ opt_jr/src/application.cpp
@@ -1,6 +1,6 @@
 #include "application.hh"
 
-Application::Application(std::string session_app_id, std::string app_id, int w,
+Application::Application(std::string session_app_id, std::string app_id, double w,
             double chi_0, double chi_C, double m, double M, double V,
             double v, double D, double csi, std::string St, int DatasetSize):
